+++ Programs/Sources/Libraries/mcon/Vectord/test_Vectord.cpp
@@ -9,8 +9,8 @@ static double _fabs(double v)
 
 void ShowVectord(const mcon::Vectord& v)
 {
-    const uint n = v.GetLength();
-    for (uint i = 0; i < n; ++i )
+    const size_t n = v.GetLength();
+    for (size_t i = 0; i < n; ++i )
     {
         printf("\t%g", v[i]);
         if ( (i % 8) == 7 )
@@ -38,11 +38,11 @@ void test_Vectord(void)
     // IsNull() is false.
     CHECK_VALUE(dvec.IsNull(), false);
 
-    for (uint i = 0; i < length; ++i)
+    for (size_t i = 0; i < length; ++i)
     {
         dvec[i] = i;
     }
-    for (uint i = 0; i < length; ++i)
+    for (size_t i = 0; i < length; ++i)
     {
         CHECK_VALUE(dvec[i], i);
     }
@@ -50,41 +50,41 @@ void test_Vectord(void)
     LOG("* [operator+=(double)]\n");
     // operator+=(T)
     dvec += 1;
-    for (uint i = 0; i < length; ++i)
+    for (size_t i = 0; i < length; ++i)
     {
         CHECK_VALUE(dvec[i], i+1);
     }
     // operator*=(T)
     LOG("* [operator*=(double)]\n");
     dvec *= 10;
-    for (uint i = 0; i < length; ++i)
+    for (size_t i = 0; i < length; ++i)
     {
         CHECK_VALUE(dvec[i], (i+1)*10);
     }
     // operator/=(T)
     LOG("* [operator/=(double)]\n");
     dvec /= 5;
-    for (uint i = 0; i < length; ++i)
+    for (size_t i = 0; i < length; ++i)
     {
         CHECK_VALUE(dvec[i], (i+1)*2);
     }
     // operator-=(T)
     LOG("* [operator-=(double)]\n");
     dvec -= 5;
-    for (uint i = 0; i < length; ++i)
+    for (size_t i = 0; i < length; ++i)
     {
         CHECK_VALUE(dvec[i], (static_cast<int>(i)+1)*2-5);
     }
     LOG("* [Copy]\n");
     mcon::Vectord dvec2(length*2);
 
-    for (uint i = 0; i < dvec2.GetLength(); ++i)
+    for (size_t i = 0; i < dvec2.GetLength(); ++i)
     {
         dvec2[i] = -(i+1);
     }
     // Copy
     dvec2.Copy(dvec);
-    for (uint i = 0; i < dvec2.GetLength(); ++i)
+    for (size_t i = 0; i < dvec2.GetLength(); ++i)
     {
         if (0 <= i && i < length)
         {
@@ -94,48 +94,48 @@ void test_Vectord(void)
     LOG("* [operator=]\n");
     // Substitution
     dvec2 = dvec;
-    for (uint i = 0; i < length; ++i)
+    for (size_t i = 0; i < length; ++i)
     {
         CHECK_VALUE(dvec2[i], (static_cast<int>(i)+1)*2-5);
     }
     dvec = 10;
-    for (uint i = 0; i < dvec2.GetLength(); ++i)
+    for (size_t i = 0; i < dvec2.GetLength(); ++i)
     {
         dvec2[i] = i + 1;
     }
 
     LOG("* [operator?=]\n");
     dvec += dvec2;
-    for (uint i = 0; i < length; ++i)
+    for (size_t i = 0; i < length; ++i)
     {
         CHECK_VALUE(dvec[i], (i+1) + 10);
     }
-    for (uint i = 0; i < dvec2.GetLength(); ++i)
+    for (size_t i = 0; i < dvec2.GetLength(); ++i)
     {
         dvec2[i] = (i & 1) ? 1.0 : 2.0;
     }
 
     dvec *= dvec2;
-    for (uint i = 0; i < length; ++i)
+    for (size_t i = 0; i < length; ++i)
     {
         CHECK_VALUE(dvec[i], ((i+1) + 10) * ((i & 1) ? 1.0 : 2.0));
     }
     dvec /= dvec2;
-    for (uint i = 0; i < length; ++i)
+    for (size_t i = 0; i < length; ++i)
     {
         CHECK_VALUE(dvec[i], (i+1) + 10);
     }
-    for (uint i = 0; i < dvec2.GetLength(); ++i)
+    for (size_t i = 0; i < dvec2.GetLength(); ++i)
     {
         dvec2[i] = i + 1;
     }
     dvec -= dvec2;
-    for (uint i = 0; i < length; ++i)
+    for (size_t i = 0; i < length; ++i)
     {
         CHECK_VALUE(dvec[i], 10);
     }
     LOG("* [Carveout]\n");
-    for (uint i = 0; i < dvec.GetLength(); ++i)
+    for (size_t i = 0; i < dvec.GetLength(); ++i)
     {
         dvec[i] = i + 1;
     }
@@ -148,13 +148,13 @@ void test_Vectord(void)
     CHECK_VALUE(dvec2.IsNull(), true);
     dvec2 = dvec(dvec.GetLength()-1, 5);
     CHECK_VALUE(dvec2.IsNull(), false);
-    for (uint i = 0; i < dvec2.GetLength(); ++i)
+    for (size_t i = 0; i < dvec2.GetLength(); ++i)
     {
         CHECK_VALUE(dvec2[i], 6);
     }
 
     dvec2 = dvec(1, 3);
-    for (uint i = 0; i < dvec2.GetLength(); ++i)
+    for (size_t i = 0; i < dvec2.GetLength(); ++i)
     {
         CHECK_VALUE(dvec2[i], i+2);
     }
@@ -162,7 +162,7 @@ void test_Vectord(void)
     LOG("* [Fifo]\n");
     double v = dvec2.Fifo(5);
     CHECK_VALUE(v, 2);
-    for (uint i = 0; i < dvec2.GetLength(); ++i)
+    for (size_t i = 0; i < dvec2.GetLength(); ++i)
     {
         CHECK_VALUE(dvec2[i], i+3);
     }
@@ -170,7 +170,7 @@ void test_Vectord(void)
     LOG("* [Unshift]\n");
     v = dvec2.Unshift(2);
     CHECK_VALUE(v, 5);
-    for (uint i = 0; i < dvec2.GetLength(); ++i)
+    for (size_t i = 0; i < dvec2.GetLength(); ++i)
     {
         CHECK_VALUE(dvec2[i], i+2);
     }
@@ -365,14 +365,14 @@ void test_Vectord(void)
         }
     }
     LOG("* [ToMatrix]\n");
-    for (uint i = 0; i < dvec.GetLength(); ++i)
+    for (size_t i = 0; i < dvec.GetLength(); ++i)
     {
         dvec[i] = i + 1;
     }
     const mcon::Matrix<double> m = dvec.ToMatrix();
     CHECK_VALUE(m.GetRowLength(), 1);
     CHECK_VALUE(m.GetColumnLength(), dvec.GetLength());
-    for (uint i = 0; i < dvec.GetLength(); ++i)
+    for (size_t i = 0; i < dvec.GetLength(); ++i)
     {
         CHECK_VALUE(m[0][i], i+1);
     }
@@ -380,7 +380,7 @@ void test_Vectord(void)
 
     CHECK_VALUE(mt.GetRowLength(), dvec.GetLength());
     CHECK_VALUE(mt.GetColumnLength(), 1);
-    for (uint i = 0; i < dvec.GetLength(); ++i)
+    for (size_t i = 0; i < dvec.GetLength(); ++i)
     {
         CHECK_VALUE(mt[i][0], i+1);
     }
